requirements.txt
sqlite3
bcrypt
pycryptodome